Yay! Algo success!

+++ TheHastyFinn/XFactor/XFactorAlgoData.cs
@@ -8,8 +8,8 @@ namespace TheHastyFinn
 {
     public class XFactorAlgoData
     {
-        public decimal HighestHi { get; set; }
-        public decimal LowestLo { get; set; }
+        public double HighestHi { get; set; }
+        public double LowestLo { get; set; }
         public int pSinceHighHi { get; set; }
         public int pSinceLowLo { get; set; }
         public int pHighHi { get; set; }
@@ -23,7 +23,7 @@ namespace TheHastyFinn
         public void ResetValues()
         {
             HighestHi = 0;
-            LowestLo = decimal.MaxValue;
+            LowestLo = double.MaxValue;
             pSinceHighHi = 0;
             pSinceLowLo = 0;
             pHighHi = 0;
